backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
python-multipart>=0.0.18
pydantic>=2.10,<3
pydantic-settings>=2.7,<3
openai>=1.75,<2
google-genai>=1.0,<2
langgraph>=0.4,<1
Pillow>=11,<12
scikit-image>=0.24,<1
httpx>=0.28,<1
asyncpg>=0.30,<1
aiolimiter>=1.1,<2
anthropic>=0.52,<1
structlog>=24.0,<25
prometheus-client>=0.21,<1
prometheus-fastapi-instrumentator>=7.0,<8
